+++ lib/seed/pre_lottery_draw_generator.rb
@@ -1,0 +1,10 @@
+# frozen_string_literal: true
+
+# Seed script generator for pre-lottery draws
+class PreLotteryDrawGenerator < DrawGenerator
+  private
+
+  def status
+    'pre_lottery'
+  end
+end